Clarified public API of grammar utils
