+++ packages/langium/src/index.ts
@@ -12,7 +12,6 @@ export * from './generator/template-string';
 export * from './grammar/ast-reflection-interpreter';
 export * from './grammar/generated/ast';
 export * from './grammar/generated/module';
-export * from './grammar/grammar-util';
 export * from './grammar/langium-grammar-module';
 export * from './grammar/language-meta-data';
 export * from './grammar/type-system/type-collector';
@@ -34,8 +33,9 @@ export * from './syntax-tree';
 export * from './utils/ast-util';
 export * from './utils/collections';
 export * from './utils/cst-util';
+export * from './utils/grammar-util';
 export * from './utils/promise-util';
-export * from './utils/uri-utils';
+export * from './utils/uri-util';
 export * from './utils/regex-util';
 export * from './utils/stream';
 export * from './validation/document-validator';
