+++ packages/langium/src/utils/uri-util.ts
