+++ packages/langium/src/workspace/ast-descriptions.ts
@@ -11,7 +11,7 @@ import { AstNode, AstNodeDescription, ReferenceInfo } from '../syntax-tree';
 import { getDocument, isLinkingError, streamAst, streamReferences } from '../utils/ast-util';
 import { toDocumentSegment } from '../utils/cst-util';
 import { interruptAndCheck } from '../utils/promise-util';
-import { equalURI } from '../utils/uri-utils';
+import { equalURI } from '../utils/uri-util';
 import { AstNodeLocator } from './ast-node-locator';
 import { DocumentSegment, LangiumDocument } from './documents';
